bug(testing) fix travis integration error [fixes #158839469]
